Create BZYtoken.sol
The code defines an ERC20-compliant token contract called "BZYtoken" with a total supply of 100,000 and 2 decimal places. It allows token transfers between addresses and approves spending of tokens by other addresses. Additionally, it includes a contract called "ApproveAndCallFallBack" that allows for the approval and execution of a function in one call.